Add files via upload
These are the methods for image reconstruction that i implemented. They are not meant to be very efficient, my priority was to understand them and implement them as they are fount in Kak's book: Principles of Computerized Tomographic Imaging. 
I might add some notes on SART since that was particularly tricky. 

I fused all of these methods in an interface. Which also has a function for detecting tumors via 
a ConvNet. This one i will upload later.